move: Validator 테스트 디렉토리 이동
