+++ src/test/java/com/kgu/studywithme/global/annotation/validation/ValidHashtagCountValidatorTest.java
@@ -1,7 +1,5 @@
-package com.kgu.studywithme.global.annotation;
+package com.kgu.studywithme.global.annotation.validation;
 
-import com.kgu.studywithme.global.annotation.validation.ValidHashtagCount;
-import com.kgu.studywithme.global.annotation.validation.ValidHashtagCountValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
