+++ src/test/java/com/kgu/studywithme/global/annotation/validation/ValidImageExtensionValidatorTest.java
@@ -1,6 +1,5 @@
-package com.kgu.studywithme.global.annotation;
+package com.kgu.studywithme.global.annotation.validation;
 
-import com.kgu.studywithme.global.annotation.validation.ValidImageExtensionValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
